Remove Python's PgpPrtPkts
Previously, pgpPrtPkts was directly exposed by the Python bindings.
Exposing such a low-level function makes it really hard to change the
interface.

But, I don't think it is actually used.  There are three reasons that
make me believe this:

First, I have searched on codesearch.debian.net and found no python
code using it.

Second, I don't think it is doing anything actually useful.
pgpPrtPkts is invoked with dig == NULL and printing == 1, so it
doesn't return the parsed data but merely prints it to stderr.  There
is no way to get the packet dump in Python.  Worse, once this function
is invoked just once, the packet parser will continue to dump packets
to stderr with no way to turning it off again.

Third, the function takes a Python Unicode object ("S"), which is
converted to UTF-8 and mustn't contain any nul bytes.  Since
pgpPrtPkts operates on binary OpenPGP data only, this function can
only dump the subset of OpenPGP data that happens to be valid UTF-8.

I think it is best to simply remove it.